[GEOT-5805] move polygon label point to geometry functions
