--- modules/unsupported/polylabel/pom.xml
@@ -1,36 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>gt-polygon-labeller</artifactId>
-	<name>polylabel</name>
-	<description>Java Port of https://github.com/mapbox/polylabel</description>
-	<dependencies>
-		<dependency>
-			<groupId>org.geotools</groupId>
-			<artifactId>gt-sample-data</artifactId>
-			<version>${project.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.geotools</groupId>
-			<artifactId>gt-process</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.geotools</groupId>
-			<artifactId>gt-geojson</artifactId>
-			<version>${project.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.geotools</groupId>
-			<artifactId>gt-main</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-	</dependencies>
-	<parent>
-		<groupId>org.geotools</groupId>
-		<artifactId>unsupported</artifactId>
-		<version>18-SNAPSHOT</version>
-	</parent>
-</project>
