--- modules/unsupported/polylabel/src/main/resources/META-INF/services/org.geotool.process.ProcessFactory
@@ -1,1 +1,0 @@
-org.geotools.polylabel.PolygonLabelProcess
