--- modules/unsupported/polylabel/src/main/resources/META-INF/services/org.geotools.filter.FunctionFactory
@@ -1,1 +1,0 @@
-org.geotools.polylabel.PolygonLabelFunctionFactory
