+++ modules/unsupported/pom.xml
@@ -173,17 +173,6 @@
 				<module>geojson</module>
 			</modules>
 		</profile>
-		<profile>
-            <id>polylabel</id>
-            <activation>
-                <property>
-                    <name>all</name>
-                </property>
-            </activation>
-            <modules>
-                <module>polylabel</module>
-            </modules>
-        </profile>
         <profile>
           <id>javafx</id>
           <modules>
@@ -214,7 +203,6 @@
 				<module>solr</module>
 				<module>css</module>
 				<module>geojsonstore</module>
-				<module>polylabel</module>
 			</modules>
 		</profile>
 	</profiles>
@@ -230,7 +218,7 @@
 		<module>geojson</module>
 		<module>mbtiles</module>
 		<module>wfs-ng</module>
-		<module>polylabel</module>
+		
 	</modules>
 
 </project>
