+++ modules/unsupported/process-geometry/src/main/java/org/geotools/process/geometry/PolygonLabelFunction.java
@@ -3,7 +3,7 @@
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
  * 
- *    (C) 2016, Open Source Geospatial Foundation (OSGeo)
+ *    (C) 2017, Open Source Geospatial Foundation (OSGeo)
  *    
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public
@@ -17,7 +17,7 @@
  *    
  */
 
-package org.geotools.polylabel;
+package org.geotools.process.geometry;
 
 import java.util.List;
