+++ modules/unsupported/process-geometry/src/main/resources/META-INF/services/org.geotools.filter.FunctionFactory
@@ -1,0 +1,1 @@
+org.geotools.process.geometry.PolygonLabelFunctionFactory
