+++ modules/unsupported/process-geometry/src/main/resources/META-INF/services/org.geotools.process.ProcessFactory
@@ -1,1 +1,2 @@
 org.geotools.process.geometry.GeometryProcessFactory
+org.geotools.process.geometry.PolygonLabelProcess
